Fix comparison of transitions, add flag to ignore update
Cost was only compared up to additive constants, i.e. the costs
x and x+2 were interpreted as equal, which is not correct in general.
Now only the higher cost x+2 is kept (worst-case).

A new flag allows to ignore the update for the comparison, as the
infinity check only considers cost and guard.