Distinguish between nontermination and unbounded runtime
Nontermination means that a loop with empty guards occurs in the program,
for which LoAT internally uses the variable INF. If this variable is present,
we now output "NO" for nontermination.

Unbounded runtime is detected in the limit problem when the cost depends on
a free variable and is output as "WORST_CASE(INF,?)" as before.