Very first attempt at analyzing nonlinear ITS problems
This is mostly a copy of the metering and strategy code for linear ITSs.
It should be possible to share large parts of the code,
this is only a preliminary implementation to obtain some first benchmarks.

accelerate_nonlinear:
Finding metering functions for nonlinear rules is a rather simple adaption
of the old code. However, some optimizations have to be reconsidered and
all of the helper method had to be ported to NonlinearRule.

strategy_nonlinear:
The basic strategy is similar to the linear case. However, the graph
structure is now much more complicated, since rules can have multiple
destinations. Hence the chaining strategies have to be adapted.
The current implementation seems to work on simple examples,
but is probably far from optimal.