strategy: Remove incoming rules that were chained with an accelerated rule
This is different from the paper and might lose relevant execution paths
in some examples, but it also removes the number of rules and avoids
rule explosion.

Essentially, this means that we assume that if there is a loop,
it is always executed (at least once).

The old implementation (for linear problems) has a heuristic when it
might be useful to keep rules instead of deleting them. Such a heuristic
might also be useful here, but for now all chained rules are deleted.