Refactor recurrence solving, fix soundness bug
The bug was in an optimization when searching the dependencyOrder.

Example: A'=B+1, B'=A+2, C'=C+1 [C < D]
Metering is straight-forward, but we find no dependency order, since A and B
depend on each other. The optimization added A==B to the guard and simplified
the remaining update to: A'=A+1, B'=A+2.

But this is not sound: Even if A==B holds before the first iteration, it does
not hold after the first iteration (for this particular example). Hence it is
not sound to simplify the update as shown above (check with e.g. A,B=1).

The fix restricts the heuristic to cases where A==B implies A'==B'.
This holds if the right-hand sides of the simplified update coincide,
e.g., it does not hold for A'=A+1, B'=A+2, but it holds for A'=A+A, B'=A+A
which we get from A'=A+B, B'=A+B (this occurs in one example).