strategy: Soundness fix in cost computation for nonlinear rules
We assume that the cost of every step is at least 1.
The old implementation added "cost >= 1" to the guard _after_ metering.
But this only ensures that the first iteration has cost >= 1, e.g.:

f(A,B) -> f(0,B-1), f(0,B-1) [B > 0] cost: A

This loop has overall cost A, since all but the first step have cost 0.
The hold implementation showed EXP (2^B) which is incorrect.
We now add "cost >= 1" _before_ metering, so it has to hold in all steps.
(which means that we cannot find a metering function for the above rule).