WIP: Refactor rule datastructure, focus on general Rule
LinearRule is now a subclass of Rule, so most of the code can
use/store objects of type Rule. LinearRule is mostly intended
as an indication that a method only accepts linear rules.

In the current design, the conversion from Rule to LinearRule
involves copying data, which is unfortunate, but probably
not a major performance concern. On the other hand, most
of the old code can easily be adapted to this design.