Rewrite chaining of accelerated rules to fix unintended chaining
In the old implementation, a chained rule could be considered
as incoming rule in a later iteration, so it could be chained again.

Although this helped for some weird examples, this was not intended
and can increase the number of rules (which is always bad).
If we want to improve on these weird examples, we should consider
to explicitly chain accelerated rules with themselves after acceleration.