chaining: We must allow self-loops when eliminating a location
Example: f -> f,g and g -> g,f.
We cannot chain this example, so we eliminate a location, say f.
Note that f -> f,g contains a self-loop at f (but not a simple loop).

We have to be careful in such cases, as the chained rule g -> g,f,g
still leads to f (which we want to eliminate). To fully eliminate f,
we thus remove f from all right-hand sides, so we get g -> g,g.