meter: Remove possibly unsound preprocessing steps
Eliminating temporary variables can be unsound if they also occur in the cost:

(1) Equality propagation also has to be applied to the cost. A cost of "free" is iterated to "free^n",
    so we assume that "free" does not change during loop execution. But if we replace "free" by
    a term like "A+B" where A or B are updated, then the cost must also be replaced to "A+B" so
    that we can correctly compute the iterated update (which is most likely not just (A+B)^n).

(2) If we eliminate a free var by replacing "A <= free && free <= B" by just "A <= B", where
    free occurs in the cost, then we again assume that the cost does not change during iteration.
    However, if A or B are updated, then free might have to change to satisfy "A <= free <= B",
    so it is not sound to eliminate "free" in this case.

To solve these problems, temporary variables ("free") are not removed in the metering code,
but should instead be removed in accelerator before metering. Note that this might not be
as effective for specific examples, since the optimization in the metering code was able
to ignore updates of non-relevant variables. This should, however, not matter for most
of the examples.