accelerate: Fix soundness of non-termination with cost 0
We now always add "cost >= 1" before metering. For linear rules,
we don't need this in general, so this might be a restriction
that can lose some examples.

We do need this to prove non-termination. Example:
  f(x,y) -{ y }> f(x+1,0)  [ x >= 0 && y >= 0 ]
This rule has cost 0 in each but the first iteration, hence
non-termination is not a sound result (it does not terminate,
but it also has bounded costs...).