accelerate: Always use the guard strengthening heuristic
This heuristic helps for updates like A := 0 or A := free.
The pre-processing for metering is now more restricted than it used to be
(in order to make sure it is sound), so this might help as a workaround.

A simple case where this can help is this one:
  f(A) -> f(free) :|: A == 0
This is clearly nonterminating when we set free/0. But free var instantiation
only looks at the guard for possible bounds (and free does not occur in the guard).
Guard strengthening adds the constraint "free == 0" to the guard,
so free var instantiation finds the bound and sets free/0, yielding nontermination.