Add flag to disable main improvements over paper
(1) Linearization (e.g. x^3 > 0 becomes z > 0 for metering)
(2) Conflict var heuristic (to handle cases like A++,B++ with A,B <= C)
(3) Constant update heuristc (cases like A := 0 or A := free)

Note that (2) and (3) are backed by the theory in the paper
(adding to the guard is always sound), but are not described.

(1) is not backed by the paper, but is relatively straight-forward
and motivated by simple examples like A++, A <= B^2.