accelerate: Avoid adding "cost >= 1" if not needed
For linear rules, this is usually not needed but may introduce nonlinear
arithmetic (so we may fail if linearization does not work).

Instead, we perform a second search for metering functions if we found
a nonterminating rule. Only in this case "cost >= 1" is important.
This case is probably rare enough to invest a second metering call.