analysis: Never output Omega(1), instead use Omega(0)
The result Omega(1) is only correct if there actually is an applicable rule
(i.e., with satisfiable guard) with cost >= 1. We currently do not check
this, so we can only say Omega(0).

A simple check would be easy (check if "guard && cost >= 1" is satisfiable
for any initial rule), but also costly (involves an smt query) and we
don't really care for constant bounds (we even remove rules with constant
costs to avoid rule explosion).