Asymptotic bugfix: We need satisfiability to prove nontermination
This is a soundness fix. A rule with cost "NONTERM" is only nonterminating
if its guard is satisfiable (a single modle suffices, but we did not check this).

Without the sat check (before this fix), LoAT incorrectly reported
nontermination in such cases. In most cases, this should not have
happened as unsat rules are removed at some points. It could however
happen for complicated rules where z3 reports unknown instead of unsat
(as we usually keep such rules, instead of deleting them).