Bugfix in preprocess: Never eliminate updated variables
Elimination by transitive closure (a <= x <= b becomes a <= b)
is only sound if x does not occur in any update. We computed
the set of suitable variables x before applying equality propagation.
But equality propagation can alter the update, so we instead have
to compute the suitable variables x after equality propagation
has already been performed.

This fixes a soundness bug, example:
  start(x) -> f(y - 1) :|: y < x && y - 1 = z && z >= 0
  f(x) -{x,x}> g(x)

The constraint "y < x" was eliminated (by transitive closure) after
replacing y by z+1. This is unsound (as z then occurs in the update)
and resulted in unbounded runtime, which is not correct.